Cleanup: timing as floats, named clamp release pin, remove Servo dep, replace magic numbers

+++ include/Config.h
@@ -31,17 +31,17 @@ extern const float RETURN_ACCEL;
 
 // Timing Settings (milliseconds)
 namespace Timing {
-extern const int CLAMP_SETTLE_TIME;
-extern const int CLAMP_RELEASE_TIME;
-extern const int HOME_SETTLE_TIME;
-extern const int MOTION_SETTLE_TIME;
-extern const int ALIGN_SETTLE_TIME;
-extern const int SIGNAL_DURATION;
-extern const int MOTOR_TIMEOUT;
+extern const float CLAMP_SETTLE_TIME;
+extern const float CLAMP_RELEASE_TIME;
+extern const float HOME_SETTLE_TIME;
+extern const float MOTION_SETTLE_TIME;
+extern const float ALIGN_SETTLE_TIME;
+extern const float SIGNAL_DURATION;
+extern const float MOTOR_TIMEOUT;
 // Alignment-specific timing (ms) to avoid magic numbers in alignment choreography
-extern const int ALIGNMENT_LEFT_CLAMP_EXTEND_MS;    // was 50
-extern const int ALIGNMENT_ALIGNMENT_MOVE_MS;       // was 20
-extern const int ALIGNMENT_RIGHT_CLAMP_WAIT_MS;     // was 150
-extern const int ALIGNMENT_SHORT_SETTLE_MS;         // was 100
-extern const int ALIGNMENT_LONG_SETTLE_MS;          // was 200
+extern const float ALIGNMENT_LEFT_CLAMP_EXTEND_MS;    // was 50
+extern const float ALIGNMENT_ALIGNMENT_MOVE_MS;       // was 20
+extern const float ALIGNMENT_RIGHT_CLAMP_WAIT_MS;     // was 150
+extern const float ALIGNMENT_SHORT_SETTLE_MS;         // was 100
+extern const float ALIGNMENT_LONG_SETTLE_MS;          // was 200
 }  // namespace Timing
