+++ src/StateMachine/STATES/Cutting Cycle/02_ALIGNMENT.cpp
@@ -70,7 +70,7 @@ void handleAlignmentState() {
             
         case 5:
             // Step 3.2: Wait (allow left clamp to fully extend)
-            if (millis() - stepStartTime >= Timing::ALIGNMENT_LEFT_CLAMP_EXTEND_MS) {
+            if ((float)(millis() - stepStartTime) >= Timing::ALIGNMENT_LEFT_CLAMP_EXTEND_MS) {
                 // Step 3.3: Extend alignment cylinder (position material precisely)
                 extendAlignmentCylinder();
                 stepStartTime = millis();
@@ -80,7 +80,7 @@ void handleAlignmentState() {
             
         case 6:
             // Step 3.4: Wait (allow alignment movement to complete)
-            if (millis() - stepStartTime >= Timing::ALIGNMENT_ALIGNMENT_MOVE_MS) {
+            if ((float)(millis() - stepStartTime) >= Timing::ALIGNMENT_ALIGNMENT_MOVE_MS) {
                 // Step 3.5: Retract left clamp (release to allow fine adjustment)
                 retractLeftClamp();
                 stepStartTime = millis();
@@ -90,7 +90,7 @@ void handleAlignmentState() {
             
         case 7:
             // Step 3.6: Wait (settle time)
-            if (millis() - stepStartTime >= Timing::ALIGNMENT_LONG_SETTLE_MS) {
+            if ((float)(millis() - stepStartTime) >= Timing::ALIGNMENT_LONG_SETTLE_MS) {
                 currentStep++;
                 stepStartTime = millis();
             }
@@ -108,7 +108,7 @@ void handleAlignmentState() {
             
         case 9:
             // Step 4.2: Wait, then retract alignment cylinder
-            if (millis() - stepStartTime >= Timing::ALIGNMENT_SHORT_SETTLE_MS) {
+            if ((float)(millis() - stepStartTime) >= Timing::ALIGNMENT_SHORT_SETTLE_MS) {
                 retractAlignmentCylinder();
                 stepStartTime = millis();
                 currentStep++;
@@ -117,7 +117,7 @@ void handleAlignmentState() {
             
         case 10:
             // Step 4.3: Wait (right clamp extension time)
-            if (millis() - stepStartTime >= Timing::ALIGNMENT_RIGHT_CLAMP_WAIT_MS) {
+            if ((float)(millis() - stepStartTime) >= Timing::ALIGNMENT_RIGHT_CLAMP_WAIT_MS) {
                 // Step 4.4: Retract right clamp (release for repositioning)
                 retractRightClamp();
                 stepStartTime = millis();
@@ -135,7 +135,7 @@ void handleAlignmentState() {
             
         case 12:
             // Step 4.6: Wait, then retract alignment cylinder (clear for cutting)
-            if (millis() - stepStartTime >= Timing::ALIGNMENT_SHORT_SETTLE_MS) {
+            if ((float)(millis() - stepStartTime) >= Timing::ALIGNMENT_SHORT_SETTLE_MS) {
                 retractAlignmentCylinder();
                 stepStartTime = millis();
                 currentStep++;
@@ -144,7 +144,7 @@ void handleAlignmentState() {
             
         case 13:
             // Step 4.8: Wait (settle time)
-            if (millis() - stepStartTime >= Timing::ALIGNMENT_LONG_SETTLE_MS) {
+            if ((float)(millis() - stepStartTime) >= Timing::ALIGNMENT_LONG_SETTLE_MS) {
                 currentStep++;
                 stepStartTime = millis();
             }
@@ -169,7 +169,7 @@ void handleAlignmentState() {
             
         case 16:
             // Wait for final clamp engagement before proceeding to cutting cycle
-            if (millis() - stepStartTime >= Timing::CLAMP_SETTLE_TIME) {
+            if ((float)(millis() - stepStartTime) >= Timing::CLAMP_SETTLE_TIME) {
                 // ADDED: Set current position to account for relative alignment moves
                 float netAlignmentSteps = (Motion::ALIGNMENT_INITIAL_DISTANCE - Motion::ALIGNMENT_BACKWARD_DISTANCE) * Motion::STEPS_PER_INCH;
                 setCurrentMotorPosition((long)netAlignmentSteps);
